config: support prefer-dark and prefer-light for window-theme
Add two new values for the window-theme option: prefer-dark and
prefer-light.  On macOS, these are identical to the dark and light
options.  On Linux, they indicate a preference for a dark or light theme
when the system dark mode setting does not explicitly indicate a
preference.  The default behavior of ghostty is to prefer light; these
new option values allow users to modify that behavior.